requirements.txt
PySide6
requests
Pillow

markdown
markdownify
orjson
